refactor: rename `mmseq2` to `seq_alignment.py` for all MSA needs
- moved MSARuner from `processors.py` to this new file
